Support updating preloaded files in a git repository
The commands in otp_build for updating the checked-in preloaded
modules only works in ClearCase.

Add support updating the preloaded files in a git repository as well.

Check the $version_controller variable so that we can have
"otp_build update_preloaded" do different things depending on
whether it is run in a ClearCase view or a git repository.

When run in a git repository, "otp_build update_preloaded" will
update and commit the preloaded modules (to undo the updating of
the preloaded modules, do "git reset --hard HEAD^").

Update the usage text to only show the commands relevant to
the kind of VCS it is being run in.